Changed how the save-scheduling works and added an auto restart feature
The save-scheduling now uses a value from the config.yml (also new) to
tell it how long between saves. I have also added and auto-restart
feature that allows the server to automatically restart after a certain
amount of time (BukkitForge or MCPC+ only), also configurable via the
config.yml, and displays a message a certain amount of time before it
happens (also configurable). I need to fix the timing slightly, and the
Log/broadcast messages need to be fixed.